whoops missed one

+++ Sine_Wave/src/waveform.h
@@ -5,7 +5,7 @@
 
 struct waveform {
     uint16_t frequency; //1-15
-    float amplitude; //1-10
+    uint16_t amplitude; //1-10
     uint16_t phase_shift; //0-15
 };
